pyproject.toml
[project]
name = "17-expense-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dateparser>=1.2.2",
    "fastapi>=0.122.0",
    "fastmcp>=2.13.1",
    "httpx>=0.28.1",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.1.14",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "prometheus-client>=0.23.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
